Stable branch with core functionality. (#1)
* Create Door class. Declare member booleans.

* Add get methods.

* Add test project. Define Door init test.

* Add Executive class. Initialize Doors.

* Move simulation out of Executive.

* Add Door accessor methods.

* Add Door accessor methods.

* Define Simulation constructor.

* Move Simulation code out of test. Begin implement.

Simulation code was incorrectly assigned to test project and was of course causing linker errors. A brief initial implementation of Simulation is also in luded in this commit.

* Initialize member vairables in initialization list.

* Add resetDoors()

* Add m_isPicked, setIsPicked().

* Add getIsPicked()

* Implement Simulation::run()

Simulation::run() will execute the required amount of simulations, however the math is not working out properly.

* Refactor Door class.

After some development time and problem analysis, I determined that I had a lot of unnecessary Door member variables. This is the first step in an attempt to pare down any unnecessary code in an effort to simplify the project.

* Refactor

Removed a lot of unnecessary member variables & methods from both Door & Simulation classes. Math is verified for the case when door is switched.

* Math works when not switching.

* Begin display implementation.

* Rewrite to access vector with []

* Implement console display.